refac: 전시 작품 네이밍 통일
