+++ components/pages/ArtworkCard/ArtworkCard.styles.tsx
