+++ components/pages/ArtworkCardList/ArtworkCardList.styles.tsx
